Fix most of the damage of the new codebase
